apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=70.2.0", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "smart_vision_api"
version = "1.0.0"
description = "FastAPI service for Smart Vision machine learning models and predictions"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "suhun.hong"}
]
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.30.0",
    "numpy>=1.26,<2",
    "pydantic_settings==2.9.1",
    "FlagEmbedding>=1.3.4",
    "transformers>=4.57.1,<5.0.0",
    "torch>=2.8.0,<2.9.0; platform_system != 'Darwin' or platform_machine != 'x86_64'",
    "pymilvus==2.5.10",
    "pymilvus[model]",
    "pymilvus[bulk_writer]",
    "pypdf>=4.2.0",
    "python-multipart==0.0.20",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "langchain-core>=0.3,<0.4",
    "langchain-openai>=0.2,<0.4"
]


[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "pytest>=8.3.5"
]

[tool.setuptools.packages.find]
include = ["smart_vision_api*"]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]
